deleted Celebrities from this branch and little change test
